pdf-editor/backend/requirements.txt
fastapi
uvicorn
pymupdf
python-multipart
Pillow
langdetect
fonttools
numpy
supabase

